deprecated `useOutput`
