+++ packages/react/README.md
@@ -1,6 +1,6 @@
 <img src="https://raw.githubusercontent.com/agile-ts/agile/master/packages/react/static/header_background.png" alt="React Integration">
 
-> Integrate AgileTs into React or React-Native
+> Integration for React and React-Native
 
  <br />
 
@@ -20,43 +20,46 @@
 <br />
 <img src="https://raw.githubusercontent.com/agile-ts/agile/master/packages/react/static/what_does_this_integration_header.png" alt="What does this Integration?"/>
 
-The `react` package helps us to integrate AgileTs into a [React](https://reactjs.org/) environment
-and serves as an Interface to React.
-Its main task is to bind States to React Components.
-This binding ensures that AgileTs rerender the Component whenever a bound State mutates.
-It also offers some other valuable functionalities that optimize the workflow using AgileTs in a React project.
+The `react` package is a set of utilities that simplifies the way AgileTs is integrated into a [React](https://reactjs.org/) environment.
+Think of it as an extension of AgileTs in the context of React
+that serves as an interface to React Components.
+The main task of the `react` integration is to bind AgileTs States to React Components.
+This ensures that AgileTs will re-render the Component when the bound State changes.
+It also provides some other valuable functionalities
+that optimize the workflow with AgileTs in a React project.
 
-A distinction is made between `Functional` and `Class` Components.
-As we prefer to use [`React Hooks`](https://reactjs.org/docs/hooks-intro.html) in Functional Components
-but Hooks aren't supported in Class Components.
-Therefore, we have created alternatives for Class Components in order to offer the same functionalities there as well.
+A distinction is made between `Functional` and `Class` React Components.
+As we prefer to use [`React Hooks`](https://reactjs.org/docs/hooks-intro.html) in Functional Components,
+however, Hooks aren't supported in Class Components.
+Therefore, we have created alternatives for Class Components
+in order to offer the most essential functionalities there as well.
 
 ### 🐆 Functional Component
-In Functional Components we recommend using AgileTs Hooks like [`useAgile()`](https://agile-ts.org/docs/react/hooks/#useagile).
+In `Functional Components` we recommend using AgileTs Hooks like [`useAgile()`](https://agile-ts.org/docs/react/hooks/#useagile).
 The `useAgile()` Hook binds [Agile Sub Instances](https://agile-ts.org/docs/introduction#agile-sub-instance) 
-(like States or Collections) to React Components.
+(like States or Collections) to React Components for reactivity.
 ```ts
-// -- myComponent.jsx ------------------------------------------
+// -- MyComponent.jsx ------------------------------------------
 
-// Binds MY_FIRST_STATE to myComponent
+// Binds MY_FIRST_STATE to 'MyComponent.jsx' for reactivity
 const myFirstState = useAgile(MY_FIRST_STATE);
 ```
 To find out more about `useAgile()`, and other Hooks provided by AgileTs,
 checkout the [AgileTs Hook documentation](https://agile-ts.org/docs/react/hooks).
 
 ### 🦖 Class Component
-For Class Components, we provide the `AgileHOC`.
+For `Class Components`, we provide the `AgileHOC`.
 The `AgileHOC` is a Higher Order Component that is wrapped around a React Component.
-It takes care of binding [Agile Sub Instances](https://agile-ts.org/docs/introduction#agile-sub-instance) 
-(like States or Collections) to the wrapped React Component.
+It takes care of binding [Agile Sub Instances](https://agile-ts.org/docs/introduction#agile-sub-instance)
+(like States or Collections) to the wrapped React Components for reactivity.
 ```ts
-// -- myComponent.jsx ------------------------------------------
+// -- MyComponent.jsx ------------------------------------------
 
-// Binds MY_FIRST_STATE to myComponent
+// Binds MY_FIRST_STATE to 'MyComponent.jsx' for reactivyty
 export default AgileHOC(myComponent, [MY_FIRST_STATE]);
 ```
-To find out more about the `AgileHOC` and AgileTs in Class Components,
-checkout the [AgileHOC documentation](https://agile-ts.org/docs/react/AgileHOC).
+To find out more about the `AgileHOC` and how to correctly use AgileTs in Class Components,
+take a look at the [AgileHOC documentation](https://agile-ts.org/docs/react/AgileHOC).
 
 
 <br />
@@ -75,7 +78,7 @@ Therefore, we have created a table that shows which versions fit together withou
 
 | @agile-ts/react  | @agile-ts/core          | NPM Version              | Supported React versions | Supports hook based components    |
 | ---------------- | ----------------------- | ------------------------ | -------------------------|---------------------------------- |
-| v0.1.1+          | v0.1.1+                 | v6+                      | 16.8+                    | Yes                               |
+| v0.1.2+          | v0.1.2+                 | v6+                      | 16.8+                    | Yes                               |
 
 _Older Versions aren't supported anymore_
