--- packages/react/src/hooks/useOutput.ts
@@ -1,66 +1,0 @@
-import {
-  Collection,
-  Group,
-  Observer,
-  State,
-  defineConfig,
-} from '@agile-ts/core';
-import {
-  AgileHookConfigInterface,
-  SubscribableAgileInstancesType,
-  useAgile,
-} from './useAgile';
-
-export function useOutput<X extends Array<SubscribableAgileInstancesType>>(
-  deps: X | [],
-  config?: AgileHookConfigInterface
-): AgileOutputHookArrayType<X>;
-
-export function useOutput<X extends SubscribableAgileInstancesType>(
-  dep: X,
-  config?: AgileHookConfigInterface
-): AgileOutputHookType<X>;
-
-export function useOutput<
-  X extends Array<SubscribableAgileInstancesType>,
-  Y extends SubscribableAgileInstancesType
->(
-  deps: X | Y,
-  config: AgileHookConfigInterface = {}
-): AgileOutputHookArrayType<X> | AgileOutputHookType<Y> {
-  return useAgile(
-    deps as any,
-    defineConfig(config, { observerType: 'output' })
-  );
-}
-
-// Array Type
-// https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-1.html
-export type AgileOutputHookArrayType<T> = {
-  [K in keyof T]: T[K] extends Collection<infer U> | Group<infer U>
-    ? U[]
-    : T[K] extends State<infer U> | Observer<infer U>
-    ? U
-    : T[K] extends undefined
-    ? undefined
-    : T[K] extends Collection<infer U> | Group<infer U> | undefined
-    ? U[] | undefined
-    : T[K] extends State<infer U> | Observer<infer U> | undefined
-    ? U | undefined
-    : never;
-};
-
-// No Array Type
-export type AgileOutputHookType<T> = T extends
-  | Collection<infer U>
-  | Group<infer U>
-  ? U[]
-  : T extends State<infer U> | Observer<infer U>
-  ? U
-  : T extends undefined
-  ? undefined
-  : T extends Collection<infer U> | Group<infer U> | undefined
-  ? U[] | undefined
-  : T extends State<infer U> | Observer<infer U> | undefined
-  ? U | undefined
-  : never;
