+++ packages/react/src/index.ts
@@ -4,7 +4,6 @@ export { useAgile } from './hooks/useAgile';
 export { AgileHOC } from './hocs/AgileHOC';
 export { useWatcher } from './hooks/useWatcher';
 export { useProxy } from './hooks/useProxy';
-export { useOutput } from './hooks/useOutput';
 export { useValue } from './hooks/useValue';
 export { useSelector } from './hooks/useSelector';
