+++ packages/react/tests/old/useAgile.spec.ts
@@ -48,6 +48,5 @@ const myCollection11 = useAgile(
 );
 
 const myGroupValue = useValue(MY_COLLECTION.getGroup('test'));
-const myGroupOutput = useOutput(MY_COLLECTION.getGroup('test'));
 const myGroupAgile = useAgile(MY_COLLECTION.getGroup('test'));
  */
